Project6/requirements.txt
# Project 06 — local / VM dependencies (export + manual loader + profiling).
# Cloud Function has its own pinned requirements at cloud_function/requirements.txt.

pymongo>=4.6
python-dotenv>=1.0
google-cloud-storage>=2.14
google-cloud-bigquery>=3.20
tqdm>=4.66
tabulate>=0.9

# Optional but useful for ad-hoc local profiling outside BigQuery
pandas>=2.2
